// react/src/types/routes.ts
export enum RootPath {
  base = '/',
  auth = '/auth',
}

export enum AppPath {
  clicker = '/clicker',
  boosters = '/boosters',
  bump = '/bump',
  friends = '/friends',
  rating = '/rating',
  tasks = '/tasks',
  vouchers = '/vouchers',
  nfts = '/nfts',
  bumpTicket = '/bump-ticket',
  debug = '/debug',
  tutorial = '/tutorial',
  announcement = '/announcement',
  userDebug = '/user-debug',
  trustWallet = '/trustwallet',
  ships = '/ships',
  airdrop = '/airdrop',
}

export enum AuthPath {
  signIn = 'signin',
  multiWallet = 'multi-wallet',
  userBlock = 'user-block',
}

export enum TasksPath {
  community = 'community',
  kols = 'kols',
  daily = 'daily',
  game = 'game',
}

export enum GamePath {
  spaceArcade = 'space-arcade',
}

export enum ShipsPath {
  banners = 'banners',
  modeSelection = 'mode-selection',
  pirate = 'pirate',
  defense = 'defense',
  defenseWidget = 'defense-widget',
  shipCraft = 'ship-craft',
}

export enum NftsPath {
  marketplace = 'marketplace',
  bumpStore = 'bump-store',
  myNfts = 'my-nfts',
  sell = 'sell',
  forSell = 'for-sell',
  shop = 'shop',
  lootbox = 'lootbox',
}

export enum LootboxPath {
  craft = 'craft',
}

export enum AirdropPath {
  seasonTasks = 'season-tasks',
  synthesisEngine = 'synthesis-engine',
  bumpFly = 'bump-fly',
  airdropRating = 'rating',
  howItWork = 'how-it-work',
}